+++ packages/ckeditor5-dev-env/lib/release-tools/utils/getpackagejson.js
@@ -5,6 +5,7 @@
 
 'use strict';
 
+const fs = require( 'fs' );
 const path = require( 'path' );
 
 /**
@@ -13,8 +14,11 @@ const path = require( 'path' );
  * This function is helpful for testing the whole process. Allows mocking the file
  * instead of create the fixtures.
  *
+ * @param {String} [cwd=process.cwd()] Where to look for package.json.
  * @returns {Object}
  */
 module.exports = function getPackageJson( cwd = process.cwd() ) {
-	return require( path.join( cwd, 'package.json' ) );
+	return JSON.parse(
+		fs.readFileSync( path.join( cwd, 'package.json' ), 'utf-8' )
+	);
 };