+++ packages/ckeditor5-dev-env/tests/release-tools/utils/getsubpackagespaths.js
@@ -8,7 +8,7 @@
 const path = require( 'path' );
 const expect = require( 'chai' ).expect;
 const sinon = require( 'sinon' );
-const mockery = require( 'mockery' );
+const proxyquire = require( 'proxyquire' );
 
 describe( 'dev-env/release-tools/utils', () => {
 	describe( 'getSubPackagesPaths()', () => {
@@ -18,30 +18,23 @@ describe( 'dev-env/release-tools/utils', () => {
 		beforeEach( () => {
 			sandbox = sinon.sandbox.create();
 
-			mockery.enable( {
-				useCleanCache: true,
-				warnOnReplace: false,
-				warnOnUnregistered: false
-			} );
-
 			getPackageJsonStub = sandbox.stub();
 			getDirectoriesStub = sandbox.stub();
 
-			mockery.registerMock( './getpackagejson', getPackageJsonStub );
-			mockery.registerMock( '@ckeditor/ckeditor5-dev-utils', {
-				tools: {
-					getDirectories: getDirectoriesStub
-				}
-			} );
-
 			sandbox.stub( path, 'join', ( ...chunks ) => chunks.join( '/' ) );
 
-			getSubPackagesPaths = require( '../../../lib/release-tools/utils/getsubpackagespaths' );
+			getSubPackagesPaths = proxyquire( '../../../lib/release-tools/utils/getsubpackagespaths', {
+				'./getpackagejson': getPackageJsonStub,
+				'@ckeditor/ckeditor5-dev-utils': {
+					tools: {
+						getDirectories: getDirectoriesStub
+					}
+				}
+			} );
 		} );
 
 		afterEach( () => {
 			sandbox.restore();
-			mockery.disable();
 		} );
 
 		it( 'returns all found packages', () => {